Projectiles,Different Approach on Weapons, ObjectPooler
Added an object pooler to the gamemanager object. The current implementation should be changed at the moment so that the gameManager handles the object pooling, not the weapon manager as it is right now (even thought that may work fine). Projectiles can now be fired from weapons, and the weapon manager handles some of the weapon set up and swtiching, as well as reloading, and firing with ammo amounts.