Health and starting UI
Added a basic Health script that can be attached to anything. This script comes with a delegate to subscribe to if you want to know when the object has its health changed. The UI is just the barebones of a player UI, contains a health bar that scales and an ammo count. Next, need to add the scripting to control these UI elements at runtime, including changing the color of the health bar as it lowers.