Normal Distribution, Projectile Physics, UI hacks
Added normal random distribution for weapons. Weapon now chooses a point to fire to using a Box-Muller Transform, using a weaponAccuracy variable that can be tweaker per weapon. Also made it so projectiles properly add force to the point they collide on objects. Also made the UI scaling better as a quick fix, will need to update later to make it not look like a mess.