Bezier Curve Movement
Added (very) basic implemenation to move an object through a bezier curve at a constant rate given a speed. Need to make it into a coroutine and tidy everything up as well as add some more checks to make sure we don't just crash when something unexpected happens. Also a way to see when we're done travelling maybe.