Better Respawning and UI fixes
Weapons now reset their ammo on player respawn. Careful, because we just recall the Start() of the weapon, so if we add anything new to it, have to make sure it doesn't screw that up OR we just make a better respawn solution (<- do that). Also made it so non server UI's are updated correctly for weapons. Found out that Commands called on clients dont set variables correclty, so had to do it in RPC's (hooray to probable future bugs because of this).