Pickups
Added health and ammo pickups, and pickup platforms that can spawn the pickup type and hold it somewhere. Reason we use the platform is so the server can spawn the pickup for all clients, but if we just place the pickup in the scene it SHOULD still work (should test).