RaycastSpawnable pt.1
Added part of a script to display a possible spawnable object for the player to decid where to spawn it. Right now it just displays the object at the hit objects location with the objects up as the hit objects normal. Need to add spawning capabilities as well as showing whether or not the object is "valid" to spawn, and rotating/mouse wheel to move it closer/farther (maybe). Also can't spawn object if the angle between the player up and the hit objects normal is bigger than a chosen float.