spawnableDisplay fixes
Fixed spawnable object displays being above/in the ground. Now calculates the correct distance our object needs to be moved up to be placed ontop of the ground exactly (or close enough). For prefabs that have a rotation, if they are spawnable (and can be displayed before spawning), then NEED to have an empty gameobject parent with no rotations or anything. This prevents having to create some complex custom algorithm for objects with weird rotations.