Spawnable Display collision check
Added the collision checking for our spawnable display object. Creates an overlap box using the objects renderer bounds and if it collides with anything, sets canSpawn to false and makes the color red instead of blue. Right now it's just an overlap box. Could use a capsule OR even add a meshcollider to our object, but this should be fine for now. Also made it so the rotation resets when the player resets the display.

Right now we're overlap boxing EVERY frame, which should be fine though becuase this is only done locally. If this does cause slowdown eventually, can go back to the meshcollider idea from above.

Need to add the actual spawning behaviour next (can start with local then move to networked) as well as actually turning this into an ability that can go on the bar/in the abilitymanager.